backend/requirements.txt
fastapi
uvicorn
opencv-python
face_recognition
Rpi.GPIO
sqlalchemy
telebot
qrcode
Pillow